Rename seed_loader.py to yaml_seed_loader.py
Make the loader's medium explicit in the filename so future non-YAML loaders (e.g. JSON, remote dataset, hub) read as siblings rather than overloading a single ''seed_loader'' module.
