+++ pyrit/models/seeds/__init__.py
@@ -20,17 +20,17 @@ from pyrit.models.seeds.seed_attack_group import SeedAttackGroup
 from pyrit.models.seeds.seed_attack_technique_group import SeedAttackTechniqueGroup
 from pyrit.models.seeds.seed_dataset import SeedDataset
 from pyrit.models.seeds.seed_group import SeedGroup
-from pyrit.models.seeds.seed_loader import (
-    load_seed_dataset_from_yaml,
-    load_seed_from_yaml,
-    load_seed_prompt_from_yaml_with_required_parameters,
-)
 from pyrit.models.seeds.seed_objective import SeedObjective
 from pyrit.models.seeds.seed_prompt import SeedPrompt
 from pyrit.models.seeds.seed_simulated_conversation import (
     NextMessageSystemPromptPaths,
     SeedSimulatedConversation,
     SimulatedTargetSystemPromptPaths,
+)
+from pyrit.models.seeds.yaml_seed_loader import (
+    load_seed_dataset_from_yaml,
+    load_seed_from_yaml,
+    load_seed_prompt_from_yaml_with_required_parameters,
 )
 
 __all__ = [
